requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.4.1
cryptography==42.0.2
Django==4.2.8
django-cors-headers==4.3.1
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==23.2.1
firebase-admin==6.4.0
google-api-core==2.17.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
msgpack==1.0.7
packaging==23.2
pluggy==1.4.0
prometheus-client==0.19.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0